chore: move `watch` command to be an option on `nargo compile`
