+++ tooling/nargo_cli/src/cli/compile_cmd.rs
@@ -1,4 +1,6 @@
+use std::io::Write;
 use std::path::Path;
+use std::time::Duration;
 
 use fm::FileManager;
 use nargo::artifacts::program::ProgramArtifact;
@@ -15,6 +17,8 @@ use noirc_frontend::graph::CrateName;
 
 use clap::Args;
 use noirc_frontend::hir::ParsedFiles;
+use notify::{EventKind, RecursiveMode, Watcher};
+use notify_debouncer_full::new_debouncer;
 
 use crate::backends::Backend;
 use crate::errors::CliError;
@@ -27,21 +31,25 @@ use rayon::prelude::*;
 /// Compile the program and its secret execution trace into ACIR format
 #[derive(Debug, Clone, Args)]
 pub(crate) struct CompileCommand {
-    /// The name of the package to compile
+    /// The name of the package to compile.
     #[clap(long, conflicts_with = "workspace")]
     package: Option<CrateName>,
 
-    /// Compile all packages in the workspace
+    /// Compile all packages in the workspace.
     #[clap(long, conflicts_with = "package")]
     workspace: bool,
 
     #[clap(flatten)]
     compile_options: CompileOptions,
+
+    /// Watch workspace and recompile on changes.
+    #[clap(long, hide = true)]
+    watch: bool,
 }
 
 pub(crate) fn run(
     backend: &Backend,
-    args: CompileCommand,
+    mut args: CompileCommand,
     config: NargoConfig,
 ) -> Result<(), CliError> {
     let toml_path = get_package_manifest(&config.program_dir)?;
@@ -54,28 +62,76 @@ pub(crate) fn run(
         selection,
         Some(NOIR_ARTIFACT_VERSION_STRING.to_owned()),
     )?;
-    let circuit_dir = workspace.target_directory_path();
 
+    if args.compile_options.expression_width.is_none() {
+        args.compile_options.expression_width = Some(backend.get_backend_info_or_default());
+    };
+
+    if args.watch {
+        watch_workspace(&workspace, &args.compile_options)
+            .map_err(|err| CliError::Generic(err.to_string()))?;
+    } else {
+        compile_workspace_full(&workspace, &args.compile_options)?;
+    }
+
+    Ok(())
+}
+
+fn watch_workspace(workspace: &Workspace, compile_options: &CompileOptions) -> notify::Result<()> {
+    let (tx, rx) = std::sync::mpsc::channel();
+
+    // No specific tickrate, max debounce time 1 seconds
+    let mut debouncer = new_debouncer(Duration::from_secs(1), None, tx)?;
+
+    // Add a path to be watched. All files and directories at that path and
+    // below will be monitored for changes.
+    debouncer.watcher().watch(&workspace.root_dir, RecursiveMode::Recursive)?;
+
+    let mut screen = std::io::stdout();
+    write!(screen, "{}", termion::cursor::Save).unwrap();
+    screen.flush().unwrap();
+    let _ = compile_workspace_full(workspace, compile_options);
+    for res in rx {
+        let debounced_events = res.map_err(|mut err| err.remove(0))?;
+
+        if debounced_events.iter().all(|event| {
+            !matches!(event.kind, EventKind::Modify(_))
+                || event
+                    .event
+                    .paths
+                    .iter()
+                    .all(|path| path.extension().map_or(false, |ext| ext != "nr"))
+        }) {
+            continue;
+        }
+        write!(screen, "{}{}", termion::cursor::Restore, termion::clear::AfterCursor).unwrap();
+        screen.flush().unwrap();
+        let _ = compile_workspace_full(workspace, compile_options);
+    }
+
+    screen.flush().unwrap();
+
+    Ok(())
+}
+
+fn compile_workspace_full(
+    workspace: &Workspace,
+    compile_options: &CompileOptions,
+) -> Result<(), CliError> {
     let mut workspace_file_manager = file_manager_with_stdlib(&workspace.root_dir);
-    insert_all_files_for_workspace_into_file_manager(&workspace, &mut workspace_file_manager);
+    insert_all_files_for_workspace_into_file_manager(workspace, &mut workspace_file_manager);
     let parsed_files = parse_all(&workspace_file_manager);
 
-    let expression_width = args
-        .compile_options
-        .expression_width
-        .unwrap_or_else(|| backend.get_backend_info_or_default());
-    let compiled_workspace = compile_workspace(
-        &workspace_file_manager,
-        &parsed_files,
-        &workspace,
-        &args.compile_options,
-    );
+    let expression_width =
+        compile_options.expression_width.expect("expression width should have been set");
+    let compiled_workspace =
+        compile_workspace(&workspace_file_manager, &parsed_files, workspace, compile_options);
 
     let (compiled_programs, compiled_contracts) = report_errors(
         compiled_workspace,
         &workspace_file_manager,
-        args.compile_options.deny_warnings,
-        args.compile_options.silence_warnings,
+        compile_options.deny_warnings,
+        compile_options.silence_warnings,
     )?;
 
     let (binary_packages, contract_packages): (Vec<_>, Vec<_>) = workspace
@@ -85,11 +141,12 @@ pub(crate) fn run(
         .partition(|package| package.is_binary());
 
     // Save build artifacts to disk.
-    let only_acir = args.compile_options.only_acir;
+    let only_acir = compile_options.only_acir;
     for (package, program) in binary_packages.into_iter().zip(compiled_programs) {
         let program = nargo::ops::transform_program(program, expression_width);
         save_program(program.clone(), &package, &workspace.target_directory_path(), only_acir);
     }
+    let circuit_dir = workspace.target_directory_path();
     for (package, contract) in contract_packages.into_iter().zip(compiled_contracts) {
         let contract = nargo::ops::transform_contract(contract, expression_width);
         save_contract(contract, &package, &circuit_dir);
