+++ tooling/nargo_cli/src/cli/mod.rs
@@ -26,7 +26,6 @@ mod new_cmd;
 mod prove_cmd;
 mod test_cmd;
 mod verify_cmd;
-mod watch_cmd;
 
 const GIT_HASH: &str = env!("GIT_COMMIT");
 const IS_DIRTY: &str = env!("GIT_DIRTY");
@@ -81,8 +80,6 @@ enum NargoCommand {
     Lsp(lsp_cmd::LspCommand),
     #[command(hide = true)]
     Dap(dap_cmd::DapCommand),
-    #[command(hide = true)] // Hidden while the feature is being built out
-    Watch(watch_cmd::WatchCommand),
 }
 
 #[cfg(not(feature = "codegen-docs"))]
@@ -126,7 +123,6 @@ pub(crate) fn start_cli() -> eyre::Result<()> {
         NargoCommand::Lsp(args) => lsp_cmd::run(&backend, args, config),
         NargoCommand::Dap(args) => dap_cmd::run(&backend, args, config),
         NargoCommand::Fmt(args) => fmt_cmd::run(args, config),
-        NargoCommand::Watch(args) => watch_cmd::run(args, config),
     }?;
 
     Ok(())
